Split up main file into different cogs. Create file for global variables.
